# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ByPass API",
        "description": "API de gestion de bypass réalisée par JOBS Conseil. Cette API permet de gérer les demandes de bypass, les équipements, les zones, les capteurs et les utilisateurs.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Serveur API ByPass"
        }
    ],
    "paths": {
...
        "/zones": {
            "get": {
                "tags": [
                    "Zones"
                ],
                "summary": "Liste des zones",
                "description": "Récupère la liste des zones avec leurs équipements",
                "operationId": "c7bcbad1f7a7a16d0d2aeb2352d5db08",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive",
                                "maintenance"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Rechercher dans le nom",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des zones paginée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Zone"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/schemas/Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Zones"
                ],
                "summary": "Créer une zone",
                "description": "Crée une nouvelle zone. Accessible uniquement aux administrateurs.",
                "operationId": "51c3a538062d2083811159b30f38400e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Zone A"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description de la zone",
                                        "nullable": true
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Bâtiment A",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive",
                                            "maintenance"
                                        ],
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Zone créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Zone"
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/schemas/Error"
                    },
                    "422": {
                        "$ref": "#/components/schemas/Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/zones/{zone}": {
            "get": {
                "tags": [
                    "Zones"
                ],
                "summary": "Détails d'une zone",
                "description": "Récupère les détails d'une zone avec ses équipements",
                "operationId": "cb4285a9029a20c76945c5bbb9a16850",
                "parameters": [
                    {
                        "name": "zone",
                        "in": "path",
                        "description": "ID de la zone",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la zone",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Zone"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/schemas/Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Zones"
                ],
                "summary": "Mettre à jour une zone",
                "description": "Met à jour les informations d'une zone. Accessible uniquement aux administrateurs.",
                "operationId": "1d5dc48ead02dad313061454bd4e32f0",
                "parameters": [
                    {
                        "name": "zone",
                        "in": "path",
                        "description": "ID de la zone",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Zone A"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "location": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive",
                                            "maintenance"
                                        ],
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Zone mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Zone"
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/schemas/Error"
                    },
                    "404": {
                        "$ref": "#/components/schemas/Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Zones"
                ],
                "summary": "Supprimer une zone",
                "description": "Supprime une zone. Accessible uniquement aux administrateurs.",
                "operationId": "f4735b595d3f15790dc60fe498acdacd",
                "parameters": [
                    {
                        "name": "zone",
                        "in": "path",
                        "description": "ID de la zone",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Zone supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Zone supprimée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/schemas/Error"
                    },
                    "404": {
                        "$ref": "#/components/schemas/Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Error": {
                "properties": {
                    "message": {
                        "description": "Message d'erreur",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Zone": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Zone A"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description de la zone",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive",
                            "maintenance"
                        ],
                        "example": "active"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Token d'authentification Bearer (obtenu via /api/auth/login)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}